Use separate counters for pre- and post-patch reboots.
`RebootCount` was used only for forcing post-patch reboots via
`RebootConfig = always`, but incremented for every reboot. As a result,
the forced post-patch reboot was skipped if pre-patch reboot was
required. This changes it to count pre-patch reboots separately.

In addition, it adds a guard against the endless patch reboot loops. The
limit was selected through common sense.

Closes: #755